requirements.txt
streamlit
groq
python-dotenv
fastapi
uvicorn
pydantic
python-multipart